drm/imx: sec-dsim_imx: correct DSIM version register access
commit 86b9841f83856b4a6d323b16ce30639345fcf1c4 from
https://source.codeaurora.org/external/imx/linux-imx

During the DSIM binding stage, only enable cfg clock is not
enough to access the version register, since at this time,
if the DSIM is still in the reset state, the register read
will always return 0. So before the version register read,
the runtime pm should be in resume state.

(This is the DSIM controller driver part change.)

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>