drm/imx: dpu: kms: Support full screen CRTC background
commit 38a3dcd3f5c0feb3f2f1e2ca5f52513cb41f9e90 from
https://source.codeaurora.org/external/imx/linux-imx

The CRTC background should be full screen instead of partial
screen, because the DRM core is likely to add configurable
background color support in the future.  We may cover the full
screen with ConstFrame0/1, upon which builds planes.  With this,
it is easier to compute each plane's layer offset vs CRTC start
point and all ConstFrame units can be controlled by CRTC.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>