MLK-23107-2 drm/imx: dpu: kms: Disable DPRC repeat_en in ->atomic_flush() where necessary
commit 6c47a4ba0e5074f350ac98d4ce6bfa0908827e3b from
https://source.codeaurora.org/external/imx/linux-imx

The SoC designer suggests us to disable DPRC repeat_en right after
FrameGen frame counter moving so that a plane with prefetch engine
can be disabled correctly on-the-fly.  This helps the plane be
enabled again later without flashing at the initial several lines
fetched by the plane.  The DPRC repeat_en disablement should be done
as soon as the FrameGen frame counter moving, which requires us to
disable preemption and irq.  However, based on tests, there is no
malfunction observed without the strict sequence implemented.
So, tag this as FIXME to achieve some simplicity for now.

Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>