LF-888: drm/bridge: adv7511: Remove support for 119M mode
commit 75d43d57b332e756c9a5d6d32adb7dd05d2b93ab from
https://source.codeaurora.org/external/imx/linux-imx

The modes requiring 119M is not a standard and tested mode by chip so, in
certain conditions, these modes are not well handled by this DSI-HDMI
converter. Since there is no detailed documentation regarding the MIPI
DPHY timings this chip supports, I'll just remove this clock from the
supported clocks.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>