MLK-23822: soc: imx: secvio: Fix boot message when nvmem not initialised
commit 5e952853c411718b90ec0e5923119e01375ee7c0 from
https://source.codeaurora.org/external/imx/linux-imx

When the nvmem subsystem is not initialised at boot, the probe
will fail and an error message will be displayed.
In this case the message should not be printed as the driver will
be probed later.

This patch checks the error code from nvmem before printing the
message.

It also fixes the cleaning path as the driver was not exiting
properly.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Reviewed-by: Iuliana Prodan <iuliana.prodan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>