ASoC: SOF: imx: Add code to manage DSP related clocks
commit 773e1bb25c6ab503ed9558bdbb19bac94abeb538 from
https://source.codeaurora.org/external/imx/linux-imx

We introduce arrays of clocks that will be enabled/disable
at resume/suspend in order to save power.

There are two types of clocks:
	* DSP IP clocks
	* DAI clocks

Normally, DAI clocks would be managed by the DSP firmware
but this is hard to achieve now, so just manage them from
Application Processor side for now.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>