LF-2539-1 ARM64: dts: imx8mq-evk-dcss-rm67199: add goodix touch support
commit b6ec33595f7b519b09c48d62b45bc6376aea0f75 from
https://source.codeaurora.org/external/imx/linux-imx

The new OLED not only use a new DSI panel(Raydium RM67199), but
also use a new touch chip, goodix gt1151q. This touch chip also
share the reset pin with the panel. According to the test, this
touch chip is not that critical with the reset operation, so
let panel control the reset pin.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>