clk: imx: blk_ctrl: Use the imx_ccm_lock to protect on runtime PM callbacks
commit 560d67e3e067e2c3b72b5f66fedd9e75e48992e7 from
https://source.codeaurora.org/external/imx/linux-imx

There is a chance the clocks might be gated while we're saving
the appropriate regs for runtime supend. There is no issue here by
using the imx_ccm_lock to also protect the reset bits. The alternative
would be to add a separate lock for the registers that hold reset bits.
That would complicate the logic since we would have to use different
locks for different registers.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>