LF-3453 rtc:imx-sc:Handle RTC interrupt during suspend/resume
commit 320db173243877c089b495a3bd564fd54a242187 from
https://source.codeaurora.org/external/imx/linux-imx

Add PM notifier to imx-sc RTC driver to handle interrupt during suspend.
If an RTC interrupt occurs when the suspend process has already started
and the RTC driver has not yet suspended, the driver will request for
the suspend to be aborted.
This patch attempts to narrow the window when RTC interrupt is missed
as wakeup event. It only fixes if the interrupt occurs after the
PM_SUSPEND_PREPAREevent is issued by the kernel.

Signed-off-by: Ranjani Vaidyanathan <ranjani.vaidyanathan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>