bus/fsl-mc: Add generic implementation for open/reset/close commands
commit fa1828a2e45deb5bc6b665746b767cb773906ab5 from
https://source.codeaurora.org/external/imx/linux-imx

The open/reset/close commands format is similar for all objects.
Currently there are multiple implementations for these commands
scattered through various drivers. The code is cavsi-identical.
Create a generic implementation for the open/reset/close commands.
One of the consumer will be the VFIO driver which needs to
be able to reset a device.

Signed-off-by: Diana Craciun <diana.craciun@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>