LF-4496 arm64: dts: imx8mp-evk: disable u1 and u2 for usb typec port
commit 7e8f2bfe14150d161dfe0d07f8450c08c2c61352 from
https://source.codeaurora.org/external/imx/linux-imx

USB device mode u1 acception is enabled by default, this is
good for power saving if there is no data transfer, if device
side wants to start transfer while at u1, it can issue link
change command to wakeup from u1 and transit to u0.

Unfortunately on imx8mp/mq, we found a problem on this wakeup
from u1, sometimes we need a very long time to transit to u0
after issue link change command, and the link change path shows
the actual link change histry is u1->u2->u0, the right link
change path should be u1->recovery->u0, this issue is still
under check with Synopsys.

Before the upstream patch commit b624b32584ac
("usb: dwc3: gadget: Fix START_TRANSFER link state check")
merged, dwc3 driver did not check the actual link state for
u1 and u2, a variable is used to record link state but never been
set to u1 or u2, so we did not issue wakeup command before
data transfer even the actual link state is at u1, the internal
hardware should be able to handle this situation and transit
to u0, so even a long time is required, we could not observe
this because there is no error log, or function break on the whole.
now with the fix commit pulled in, we start to check the real
link state before every data transfer and issue a wakeup command
if it is not at u0, if the wakeup can not complete in a polling
of register read 20000 times, dwc3 driver will throw a warninng
to user and continue data transfer anyway, so this underlying
issue is exposed by this console log.

This wakeup timeout issue also brings a side effect, as the
20000 times register polling is in a spinlock, which will impact
other module iterrupt handling, considering this, we disable the
u1 and u2 acception for device mode as a temporary solution before
this issued is resovled.

Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>