LF-4506-2: drm: bridge: cdns_hdmi: save new connector state
commit 41f7e5b8998840c0046114e2bcc22bcd175ce8e3 from
https://source.codeaurora.org/external/imx/linux-imx

Colorspace connector property is use to choose the right output colorspace
in mode_fixup.
The new colorspace state is not stored to the current state pointer until in atomic_commit.
The current state pointer is not updated and incorrect output colorspace
will be selected in mode_fixup.

Save new connector state to struct cdns_mhdp_device and use the new
state to select correct output colorspace.

Signed-off-by: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>