LF-4428-1: ASoC: rpmsg_wm8960: Fix clock configuration on slave mode
commit fc1476ecb8f5d9933560815e9989418b074af7c1 from
https://source.codeaurora.org/external/imx/linux-imx

There is a noise issue for 8kHz sample rate on slave mode.
Compared with master mode, the difference is the DACDIV
setting, after correcting the DACDIV, the noise is gone.

There is no noise issue for 48kHz sample rate, because
the default value of DACDIV is correct for 48kHz.

So wm8960_configure_clocking() should be functional for
ADC and DAC function even if it is slave mode.

In order to be compatible for old use case, just add
condition for checking that sysclk is zero with
slave mode.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Peng Zhang <peng.zhang_8@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>