dpaa2-mac: get rid of hackish detach/attach functions
commit 964e0a8bfceb3a6629bfd92bc39858de7a0c7a92 from
https://source.codeaurora.org/external/imx/linux-imx

If we want the DPAA2 MAC support to be capable of building as a module
integrated into a driver (dpaa2-eth, dpaa2-switch etc) then we cannot
use hacking API for detaching/attaching a driver to the DPMAC object.

Remove the hackish include file and use generic functions for this
purpose.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>